refacto to split cmd and functions
